🔨refactor: common 폴더로 ProfileList 이동
